fix(auto-router): clear frontend-lint and type-discipline gates, trim LOC
